mysite/requirements.txt
Django>=3.0,<=3.1
whitenoise
psycopg2-binary
python-dateutil
